backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
defusedxml==0.7.1
Django==4.2.19
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.15.2
djoser==2.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.2.0
fpdf==1.7.2
gunicorn==20.1.0
h11==0.14.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
